refactor(server): make serve() a library entrypoint; move workers into extenddb-server
The server orchestration (component assembly, cache wiring, AppState
construction, worker spawning, TLS assembly, PID-file cleanup) lived in the bin
crate, so a third party holding only their backend crate plus the published
extenddb-* library crates could not run a server without forking bin.

Move that orchestration into extenddb-server as a public `serve(config,
listener, port, run_dir, foreground, git_hash)` entrypoint, and move the
generic background workers (log-level poll, throttling poll, metrics
prune/flush, login-attempt cleanup, capacity warning) into the server crate
alongside it. AppState/start_server/Router/caches already lived here, so this
completes the assembly into one library call.

Build provenance (git hash) is passed in by the caller rather than read via
env!(): the deployed binary knows its own provenance, and the library must not
depend on the bin's build.rs environment variables. log_to_syslog_raw moves
into the server crate with serve; the bin retains only the CLI concerns
(config permission check, arg parsing, banner, bind, daemonize, PID dir).

bin's cmd_serve::run now loads config, binds, daemonizes, then calls
extenddb_server::serve. The unused syslog-tracing dependency is dropped from
bin.

Behavior-preserving: workspace unit tests green, clippy -D warnings + fmt clean.
