+++ crates/bin/Cargo.toml
@@ -35,7 +35,6 @@ time = { workspace = true }
 toml = { workspace = true }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
-syslog-tracing = { workspace = true }
 base64 = { workspace = true }
 libc = { workspace = true }
 rcgen = { workspace = true }
